+++ tianshou/policy/__init__.py
@@ -23,7 +23,7 @@ from tianshou.policy.modelfree.redq import REDQ
 from tianshou.policy.modelfree.discrete_sac import DiscreteSAC
 from tianshou.policy.imitation.base import ImitationLearning
 from tianshou.policy.imitation.bcq import BCQ
-from tianshou.policy.imitation.cql import CQLPolicy
+from tianshou.policy.imitation.cql import CQL
 from tianshou.policy.imitation.td3_bc import TD3BCPolicy
 from tianshou.policy.imitation.discrete_bcq import DiscreteBCQ
 from tianshou.policy.imitation.discrete_cql import DiscreteCQLPolicy
@@ -55,7 +55,7 @@ __all__ = [
     "DiscreteSAC",
     "ImitationLearning",
     "BCQ",
-    "CQLPolicy",
+    "CQL",
     "TD3BCPolicy",
     "DiscreteBCQ",
     "DiscreteCQLPolicy",